requirements.txt
django==5.1.3
djangorestframework==3.15.2
psycopg==3.2.3
drf-spectacular==0.27.2